__OTHER_FRAMEWORKS__/open-telemetry/.project-main/migrate-otel-projects/pyproject.toml
[project]
name = "migrate-otel-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "llm>=0.26",
    "markdown>=3.8.2",
    "pydantic>=2.11.7",
    "pygithub>=2.6.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.0",
]
